Start refactoring oh boy
