+++ internal/system/cmd.go
@@ -1,0 +1,5 @@
+package system
+
+// TODO HERE PRIVATE ONLY
+type cmd interface {
+}
