+++ internal/system/filesystem.go
@@ -1,4 +1,4 @@
-package files
+package system
 
 import "os"
